refactor: rewrite client startup procedure & remove proxy mode
